[clang] Move SingleWarningStackAwareExecutor to a separate file.
