+++ clang/include/clang/Basic/Stack.h
@@ -16,8 +16,6 @@
 
 #include <cstddef>
 
-#include "clang/Basic/Diagnostic.h"
-#include "clang/Basic/SourceLocation.h"
 #include "llvm/ADT/STLExtras.h"
 #include "llvm/Support/Compiler.h"
 
@@ -52,31 +50,6 @@ namespace clang {
     Fn();
 #endif
   }
-
-  class SingleWarningStackAwareExecutor {
-  public:
-    SingleWarningStackAwareExecutor(DiagnosticsEngine &diags)
-        : DiagsRef(diags) {}
-
-    /// Run some code with "sufficient" stack space. (Currently, at least 256K
-    /// is guaranteed). Produces a warning if we're low on stack space and
-    /// allocates more in that case. Use this in code that may recurse deeply to
-    /// avoid stack overflow.
-    void runWithSufficientStackSpace(SourceLocation Loc,
-                                     llvm::function_ref<void()> Fn);
-
-    /// Check to see if we're low on stack space and produce a warning if we're
-    /// low on stack space (Currently, at least 256Kis guaranteed).
-    void warnOnStackNearlyExhausted(SourceLocation Loc);
-
-  private:
-    /// Warn that the stack is nearly exhausted.
-    void warnStackExhausted(SourceLocation Loc);
-
-    DiagnosticsEngine &DiagsRef;
-    bool WarnedStackExhausted = false;
-  };
-
 } // end namespace clang
 
 #endif // LLVM_CLANG_BASIC_STACK_H
