+++ clang/lib/CodeGen/CodeGenModule.h
@@ -26,7 +26,7 @@
 #include "clang/Basic/LangOptions.h"
 #include "clang/Basic/NoSanitizeList.h"
 #include "clang/Basic/ProfileList.h"
-#include "clang/Basic/Stack.h"
+#include "clang/Basic/SingleWarningStackAwareExecutor.h"
 #include "clang/Basic/TargetInfo.h"
 #include "clang/Basic/XRayLists.h"
 #include "clang/Lex/PreprocessorOptions.h"
